code/requirements.txt
bounding_box==0.1.3
librosa==0.8.1
masks==0.1.3
matplotlib==3.0.3
numpy==1.19.1
opencv_python==4.5.3.56
scikit_learn==1.3.1
torch==1.10.1
arguments==76
datasets==2.14.5
numpy==1.19.1
pandas==0.24.1
Pillow==10.1.0
pydub==0.25.1
scipy==1.5.3
SoundFile==0.10.3.post1
torch==1.10.1
torchaudio==0.10.1
torchvision==0.9.0
utils==1.0.1
torchgeometry==0.1.2
timm==0.4.5
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
moviepy==1.0.3